提取请求url中文解析为单个中间件
