+++ app.js
@@ -29,6 +29,7 @@ const comStr  = require('./bin/commandStr')
 const _command = comStr.commandStr();
 const open    = require('./bin/open');
 const rotues  = require('./bin/rotues');
+const parseURL  = require('./bin/parseURL');
 
 // 系统配置
 const iservers = require('./config');
@@ -81,49 +82,7 @@ if (iservers.type === 'SERVER') {
 	
 }
 
-app.use((req, res, next)=>{
-	let decodeURL = '';
-	
-	try {
-		decodeURL = decodeURIComponent(req.originalUrl)
-	} catch(e) {
-		
-		let convertUrl = (url)=> {
-			// 匹配 GBK 编码内容
-			let reg = /(\%\S\S)+/gi;
-
-			let decodeStr = (str)=> {
-				let arr = str.split('%');
-				arr.shift();
-				// 生成 Buffer 点位
-				let buf = new Buffer(arr.length);
-				// 为 Buffer 赋值
-				arr.forEach((hex, i) => {
-					let v = parseInt(hex, 16);
-					buf[i] = v;
-				});
-
-				// 解析
-				return iconv.decode(buf, 'gbk');
-			};
-			
-			// 匹配出 GBK 内容
-			let result = url.match(reg).sort().reverse();
-
-			result.forEach(function(str){
-				url = url.replace(str, decodeStr(str));
-			});
-
-			return url
-		}
-
-		decodeURL = convertUrl(req.path, 'gbk');
-	}
-
-	req.url= req.originalUrl = decodeURL;
-
-	next()
-})
+app.use(parseURL)
 
 // 使用路由
 rotues(app, iservers.type);
